+++ src/core/objective.jl
@@ -54,8 +54,15 @@ function objective_min_shed_load_block_rolling_horizon(pm::AbstractUnbalancedPow
     end
 
     obj_opts = Dict(n=>ref(pm, n, :options, "objective") for n in nw_ids(pm))
+    no_weights = first(obj_opts).second["disable-load-block-weight-cost"]
 
-    if first(obj_opts).second["disable-load-block-weight-cost"]
+    load_weights = Dict(
+        n => Dict(
+            l => no_weights ? 1.0 : ref(pm, n, :block_weights, b) / length(ref(pm, n, :block_loads, b)) for b in ids(pm, n, :blocks) for l in ref(pm, n, :block_loads, b)
+        ) for n in nw_ids(pm)
+    )
+
+    if no_weights
         block_weights = Dict(n => Dict(i => 1.0 for i in ids(pm, n, :blocks)) for n in nw_ids(pm))
     else
         block_weights = Dict(n => ref(pm, n, :block_weights) for n in nw_ids(pm))
@@ -64,6 +71,7 @@ function objective_min_shed_load_block_rolling_horizon(pm::AbstractUnbalancedPow
     JuMP.@objective(pm.model, Min,
         sum(
             sum( block_weights[n][i] * Int(!obj_opts[n]["disable-load-block-shed-cost"]) * (1-var(pm, n, :z_block, i)) for (i,block) in nw_ref[:blocks])
+            + sum(load_weights[n][i] * Int(!obj_opts[n]["disable-load-block-shed-cost"]) * (1-var(pm, n, :z_demand, i)) for (i,load) in nw_ref[:dispatchable_loads])
             + sum( Int( obj_opts[n]["enable-switch-state-open-cost"]) * ref(pm, n, :switch_scores, l)*(1-var(pm, n, :switch_state, l)) for l in ids(pm, n, :switch_dispatchable) )
             + sum( Int(!obj_opts[n]["disable-switch-state-change-cost"]) * sum(var(pm, n, :delta_sw_state, l)) for l in ids(pm, n, :switch_dispatchable)) / n_dispatchable_switches[n]
             + sum( Int(!obj_opts[n]["disable-storage-discharge-cost"]) * (strg["energy_rating"] - var(pm, n, :se, i)) for (i,strg) in nw_ref[:storage]) / total_energy_ub
@@ -208,7 +216,15 @@ function objective_min_shed_load_block(pm::AbstractUnbalancedPowerModel)
 
     obj_opts = Dict(n=>ref(pm, n, :options, "objective") for n in nw_ids(pm))
 
-    if first(obj_opts).second["disable-load-block-weight-cost"]
+    no_weights = first(obj_opts).second["disable-load-block-weight-cost"]
+
+    load_weights = Dict(
+        n => Dict(
+            l => no_weights ? 1.0 : ref(pm, n, :block_weights, b) / length(ref(pm, n, :block_loads, b)) for b in ids(pm, n, :blocks) for l in ref(pm, n, :block_loads, b)
+        ) for n in nw_ids(pm)
+    )
+
+    if no_weights
         block_weights = Dict(n => Dict(i => 1.0 for i in ids(pm, n, :blocks)) for n in nw_ids(pm))
     else
         block_weights = Dict(n => ref(pm, n, :block_weights) for n in nw_ids(pm))
@@ -217,6 +233,7 @@ function objective_min_shed_load_block(pm::AbstractUnbalancedPowerModel)
     JuMP.@objective(pm.model, Min,
         sum(
             sum( block_weights[n][i] * Int(!obj_opts[n]["disable-load-block-shed-cost"]) * (1-var(pm, n, :z_block, i)) for (i,block) in nw_ref[:blocks])
+            + sum(load_weights[n][i] * Int(!obj_opts[n]["disable-load-block-shed-cost"]) * (1-var(pm, n, :z_demand, i)) for (i,load) in nw_ref[:dispatchable_loads])
             + sum( Int(obj_opts[n]["enable-switch-state-open-cost"]) * ref(pm, n, :switch_scores, l)*(1-var(pm, n, :switch_state, l)) for l in ids(pm, n, :switch_dispatchable) )
             + sum( Int(!obj_opts[n]["disable-switch-state-change-cost"]) * sum(var(pm, n, :delta_sw_state, l)) for l in ids(pm, n, :switch_dispatchable)) / n_dispatchable_switches[n]
             + sum( Int(!obj_opts[n]["disable-storage-discharge-cost"]) * (strg["energy_rating"] - var(pm, n, :se, i)) for (i,strg) in nw_ref[:storage]) / total_energy_ub